refactor(doctor): split god file into command/report/view layers
doctor.go had grown to ~860 lines mixing three concerns. Split into:
  - doctor.go        command wiring + run orchestration
  - doctor_report.go report data model + gathering logic (service)
  - doctor_view.go   TUI list items + non-TTY static rendering (presentation)

Pure code move within package main; no behavior change. JSON and static
output verified byte-identical via smoke tests.
